♻️ refactor(high_level): centralize cache folder and font assets handling
- move create_cache_folder and download_font_assets to high_level module
- streamline font asset downloading logic
- improve cache folder creation with error handling

♻️ refactor(main): delegate initialization to high_level module

- remove duplicate create_cache_folder and download_font_assets functions
- call high_level.init from main and cli functions to centralize initialization
